[Refactoring]Move codes about gop
